osint-processor/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
elasticsearch==8.11.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
schedule==1.2.0
celery==5.3.4
python-dateutil==2.8.2
lxml==4.9.3
newspaper3k==0.2.8
yake==0.4.8
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
